# docs/SetupDocs/nextxr_openapi.yaml
openapi: 3.0.3
info:
  title: NextXR Digital Twin Platform API
  description: |
    Ontology-driven industrial digital-twin API. Covers twin lifecycle, the
    semantic graph and its governance, the machine-twin physics runtime,
    telemetry ingestion and history, field-protocol connectors, solar PV
    analytics, the embedded AI copilot, agentic twin authoring, 3-D
    reconstruction, the live event bus and platform health.

    136 endpoints across 12 routers plus a mounted 3-D sub-application.

    ## Authentication

    Two credential types, deliberately not interchangeable.

    - **`X-API-Key`** — a human/service credential carrying a **role**
      (`admin`, `write`, `read`) and a **tenant scope**. Configured as a JSON
      array in `NXR_API_KEYS`; the built-in demo keys load **only** when that
      variable is unset (i.e. never in a correctly-configured deployment).
      Set `NXR_REQUIRE_AUTH=1` to reject keyless calls even before keys load.
    - **`X-Device-Token`** — a telemetry device credential granting exactly one
      verb (append telemetry) in exactly one tenant, optionally narrowed to an
      asset prefix. Returned once at registration; only the hash is stored.
      It is never accepted on a read, control or copilot route, and an API key
      is never accepted on the ingest routes in its place.

    > While `NXR_API_KEYS` is unset the API is **OPEN** — every `/api` call,
    > including the LLM-billing `/copilot` endpoints, is served without a key.

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/entities/{node_id}/history:
    get:
      tags: [historian]
      summary: Measurement history for an entity over a time range
      description: |
        Serves the smallest sufficient source (raw, then 1m, 1h, 1d). Means are
        computed from `count` and `sum_value` rather than stored as `avg`,
        because an average cannot be re-averaged.
      operationId: entityHistory
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/TenantQuery'
        - name: signal
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: resolution
          in: query
          schema: { type: string, enum: [raw, 1m, 1h, 1d] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HistoryPoint' }
          description: Time series
  /api/v1/entities/{node_id}/latest:
    get:
      tags: [historian]
      summary: Latest value per signal for an entity
      operationId: entityLatest
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/TenantQuery'
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/twins/{tenant}/signals:
    get:
      tags: [historian]
      summary: Signal catalogue with last-seen times
      description: Backs the tag-mapping UI and the staleness check.
      operationId: twinSignals
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonArray' }
  /api/v1/twins/{tenant}/trends:
    get:
      tags: [historian]
      summary: Aligned multi-signal series for charting
      operationId: twinTrends
      parameters:
        - $ref: '#/components/parameters/TenantPath'
        - name: signals
          in: query
          schema: { type: string }
          description: Comma-separated signal names
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: bucket
          in: query
          schema: { type: string, enum: [1m, 1h, 1d] }
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/twins/{tenant}/history/stats:
    get:
      tags: [historian]
      summary: Coverage, backend and data quality
      operationId: twinHistoryStats
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200':
          description: |
            Includes `backend` (`timescale` / `postgres` / `sqlite`) and
            `scale_safe` — the field to alarm on. A deployment that declared
            Timescale mandatory and did not get it is not healthy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HistorianInfo' }

  # ==========================================================================
  # connectors
  # ==========================================================================
...
  /api/v1/predict:
    post:
      tags: [hub]
      summary: Hub-facing prediction surface
      operationId: hubPredict
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/assets/{asset_id}/ar-overlay:
    get:
      tags: [hub]
      summary: AR overlay payload for an asset
      operationId: hubArOverlay
      parameters:
        - name: asset_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/TenantQuery'
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }

  # ==========================================================================
  # feed
  # ==========================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Carries a role (`admin` / `write` / `read`) and a tenant scope. Configured
        via `NXR_API_KEYS`.
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: |
        A telemetry device credential: one verb (append), one tenant, optionally
        one asset prefix. Returned once at registration; only the hash is stored.
  parameters:
    TenantQuery:
      name: tenant
      in: query
      required: true
      schema: { type: string }
      description: Tenant id; must be inside the caller's key scope
    TenantPath:
      name: tenant
      in: path
      required: true
      schema: { type: string }
      description: Tenant id; checked against the caller's key scope after routing
    NodeId:
      name: node_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    JsonObject:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    JsonArray:
      description: Success
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
...
  schemas:
...
    HistorianInfo:
      type: object
      properties:
        backend: { type: string, enum: [timescale, postgres, sqlite] }
        timescale_version: { type: string }
        required: { type: boolean }
        scale_safe: { type: boolean }
        samples: { type: integer }
        earliest: { type: string, format: date-time }
        latest: { type: string, format: date-time }
        bad_quality_share: { type: number }
...
    HistoryPoint:
      type: object
      properties:
        ts: { type: string, format: date-time }
        value: { type: number, nullable: true }
        count: { type: integer, description: Present on rollup resolutions }
        min: { type: number }
        max: { type: number }
        bad_count: { type: integer }
        unit: { type: string }
...